event check in and club sponsors added
sponsor information is now displayed on the specific club pages, closing #70
students can now check in to events when it is a location based event. points will be added to the account. this closes #73
also learned that main activity does not update points upon returning to the page, so this will need to be fixed