Adds Backend Support for Fetching Event Reqs
In the admin api, I created routes for general fetching and fetching a specific request.
They are separate because not all information needs to be displayed in the list view.
The next step is to take the information from a specific request and allow the admin user to translate it to an event.

Work for #74