Feature - Adds Profile page and API Support
Adds profile page accessible through the upper right button on most root
views. Presents a basic list of the Users profile information including
but not limited to Name, StudentID, Points, Grade, and House. Needs
further development to fix the stylistic choices and display the
information in a more pleasant manner.