Starts adding AllowedGrades, Fixes bug
For the frontend for rewards, there is an allowedGrades thats gets sent to/from backend encoded as bits.
The dropdown select menu is not outputting items, and I'm not 100% sure why. Also needs to add the starting value for editing a reward.
Fixed a bug with the event request backend.

Works on #83 and #82